blood-donation-backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aniso8601==7.0.0
    # via graphene
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
beautifulsoup4==4.13.4
    # via bs4
bs4==0.0.1
    # via -r requirements.in
build==1.2.2.post1
    # via pip-tools
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via pip-tools
django==3.2.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-graphql-jwt
    #   django-redis
    #   graphene-django
django-cors-headers==4.5.0
    # via -r requirements.in
django-extensions==3.0.9
    # via -r requirements.in
django-graphql-jwt==0.3.1
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
graphene==2.1.9
    # via graphene-django
graphene-django==2.16.0
    # via django-graphql-jwt
graphql-core==2.3.2
    # via
    #   django-graphql-jwt
    #   graphene
    #   graphene-django
    #   graphql-relay
graphql-relay==2.0.1
    # via graphene
idna==3.10
    # via requests
packaging==25.0
    # via build
pip-tools==7.4.1
    # via -r requirements.in
promise==2.3
    # via
    #   graphene-django
    #   graphql-core
    #   graphql-relay
psycopg2==2.9.10
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyjwt==1.7.0
    # via
    #   -r requirements.in
    #   django-graphql-jwt
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytz==2025.2
    # via django
redis==6.2.0
    # via django-redis
requests==2.32.4
    # via -r requirements.in
rx==1.6.3
    # via graphql-core
selenium==3.141.0
    # via -r requirements.in
singledispatch==4.1.2
    # via graphene-django
six==1.17.0
    # via
    #   graphene
    #   graphql-core
    #   graphql-relay
    #   promise
soupsieve==2.7
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
text-unidecode==1.3
    # via graphene-django
typing-extensions==4.14.0
    # via beautifulsoup4
urllib3==2.4.0
    # via
    #   requests
    #   selenium
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
